Simplify testcase
It does not segfault anymore, but the problem is the same
and the verifier complains.

If the upper 32-bit of a 64-bit shift operation are not used,
then the lower shift has a dependency edge to an unused node.
The scheduler never schedules this node, because he never
encounters the upper operation.

[r28432]