Add test to demonstrate broken Mux handling design.
During if conversion, we assure that we can lower "ConvHu(Bs) > 0 ? 1 : 0"
to IR operations (Minus(ConvHu(Bs)) >> 15) and thus the if conversion
creates a corresponding Mux node. The conversion to IR operations should
happen during the localopts. However, the localopts first optimizes the
comparison: Mux(Bs != 0, 0, 1). Afterwards, we are not able to lower the
Mux to IR operations anymore. Finally, some backends die, because they
cannot transform the Mux node.